Fix unknown type parameter during CA generation (#15401)
* Fix unknown type parameter during CA generation

When generating a new PKI CA from the Web UI, the UI incorrectly sends
the type parameter in the POST body. The server will now warn on unknown
parameters, resulting in the UI surfacing that up to the caller. Since
the type is part of the URL, we don't need to duplicate it in the POST
body, so elide it.

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>

* Update ui/app/adapters/pki-ca-certificate.js

Co-authored-by: claire bontempo <68122737+hellobontempo@users.noreply.github.com>

Co-authored-by: claire bontempo <68122737+hellobontempo@users.noreply.github.com>