When database plugin Clean is called, close connections in goroutines
We are seeing long pre-seal times when nodes are being shut down, which
seems to be delays in releasing the global stateLock, which further
seems to be delayed by cleaning up the builtin database plugin.

The database plugin `Clean()` eventually calls, synchronously, `Close()`
on every database instance that has been loaded. The database instance
`Close()` calls can take a long time to complete (e.g.,
[mongo](https://github.com/hashicorp/vault/blob/v1.10.3/plugins/database/mongodb/connection_producer.go#L132)
can take up to a minute to close its client before it will return).

So, instead, we change the `Clean()` method to close all of the database
instances in goroutines so they will be closed more quickly and won't
block the release of the global stateLock. This should be safe because
`Clean()` will only be called when the plugin is being unloaded, and so
the connections will no longer be used.

In addition, this adds metrics tracking how many of each type of
databases are being used by the builtin database plugin.