[QTI-295] Add Autopilot upgrade scenario
Add `autopilot` scenario that works with Vault Enterprise editions that
are using the integrated Raft storage backend. The scenario has many
variants for changing the architecture or distro, but the primary
differences in behavior relate to the Vault `seal` variant. The scenario
currently supports either `shamir` or `awskms` seal types.

The scenario works by first provisioning a cluster with Vault Enterprise at
version 1.11.0, provisioning new nodes with the candidate version of
Vault, and then allowing auto-pilot to handle upgrading.

The initial version is configurable but Vault 1.11.1 and 1.11.2 have issues
with auto-unseal so they cannot be used with the `awskms` variants.
Shamir should support any version in the 1.11.x release line.

Signed-off-by: Ryan Cragun <me@ryan.ec>
Co-authored-by: Jaymala Sinha <jaymala@hashicorp.com>